requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
banks==2.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
llama-cloud==0.1.44
llama-cloud-services==0.6.79
llama-index-core==0.14.7
llama-index-instrumentation==0.4.2
llama-index-workflows==2.11.1
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
openai==2.7.1
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
wrapt==2.0.1
yarl==1.22.0
